Improve urllib error handling
Fixes a defect where we catch urllib.error.URLError, and then introduce
special handling for client errors based on the `status` attribute.
URLError is a base class that is only guaranteed to have a `reason`
attribute.  Instead we should be checking the subclass `HTTPError`, and
using the `code` attribute to determine the status code.